requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
channels==4.0.0
daphne==4.0.0
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
django-extensions==3.2.3
django-simple-history==3.4.0
django-phonenumber-field==7.2.0
phonenumbers==8.13.26
